requirements.txt
datasets==3.6.0
numpy==2.2.5
Pillow==11.2.1
scikit_learn==1.6.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
